services/ml-export-engine/requirements.txt
# VALUE-UP AI 백엔드 의존성

# 웹 프레임워크
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# 머신러닝
scikit-learn>=1.3.0
xgboost>=2.0.0
shap>=0.44.0

# 모델 저장/로드
joblib>=1.3.0

# 유틸리티
python-multipart>=0.0.6
requests>=2.31.0
python-dotenv>=1.0.0

# Redis 캐시
redis>=5.0.0

# 모니터링
prometheus-client>=0.19.0

# 스케줄링
schedule>=1.2.0

# 로깅
python-json-logger>=2.0.7
